Refactor constants to remove circular import
