+++ docker/.env.default
@@ -3,6 +3,9 @@ PUBLIC_DATA_PATH=./girder_data
 GIRDER_ADMIN_USER=admin
 GIRDER_ADMIN_PASS=letmein
 
+# Production time zone for backup and automated deploy
+# TIMEZONE='America/New_York'
+
 # Leave empty to use local directory
 # DIVE_GIRDER_ASSETSTORE=/opt/girder/assetstore
 # DIVE_GIRDER_MONGO_DATA=/opt/girder/mongo
